Updating Selenium and chardet
Reference #1885

+++ src/bin/selenium/webdriver/firefox/webdriver.py
@@ -28,15 +28,16 @@ except NameError:  # Python 3.x
 import shutil
 import socket
 import sys
-import types
+from contextlib import contextmanager
+
+from selenium.webdriver.common.desired_capabilities import DesiredCapabilities
+from selenium.webdriver.remote.webdriver import WebDriver as RemoteWebDriver
 
 from .extension_connection import ExtensionConnection
 from .firefox_binary import FirefoxBinary
 from .firefox_profile import FirefoxProfile
 from .options import Options
 from .remote_connection import FirefoxRemoteConnection
-from selenium.webdriver.common.desired_capabilities import DesiredCapabilities
-from selenium.webdriver.remote.webdriver import WebDriver as RemoteWebDriver
 from .service import Service
 from .webelement import FirefoxWebElement
 
@@ -45,6 +46,9 @@ class WebDriver(RemoteWebDriver):
 
     # There is no native event support on Mac
     NATIVE_EVENTS_ALLOWED = sys.platform != "darwin"
+
+    CONTEXT_CHROME = "chrome"
+    CONTEXT_CONTENT = "content"
 
     _web_element_cls = FirefoxWebElement
 
@@ -99,6 +103,7 @@ class WebDriver(RemoteWebDriver):
         """
         self.binary = None
         self.profile = None
+        self.service = None
 
         if capabilities is None:
             capabilities = DesiredCapabilities.FIREFOX.copy()
@@ -130,7 +135,9 @@ class WebDriver(RemoteWebDriver):
 
         # W3C remote
         # TODO(ato): Perform conformance negotiation
+
         if capabilities.get("marionette"):
+            capabilities.pop("marionette")
             self.service = Service(executable_path, log_path=log_path)
             self.service.start()
 
@@ -176,16 +183,19 @@ class WebDriver(RemoteWebDriver):
             # Happens if Firefox shutsdown before we've read the response from
             # the socket.
             pass
-        if "specificationLevel" in self.capabilities:
+
+        if self.w3c:
             self.service.stop()
         else:
             self.binary.kill()
-        try:
-            shutil.rmtree(self.profile.path)
-            if self.profile.tempfolder is not None:
-                shutil.rmtree(self.profile.tempfolder)
-        except Exception as e:
-            print(str(e))
+
+        if self.profile is not None:
+            try:
+                shutil.rmtree(self.profile.path)
+                if self.profile.tempfolder is not None:
+                    shutil.rmtree(self.profile.tempfolder)
+            except Exception as e:
+                print(str(e))
 
     @property
     def firefox_profile(self):
@@ -195,3 +205,25 @@ class WebDriver(RemoteWebDriver):
 
     def set_context(self, context):
         self.execute("SET_CONTEXT", {"context": context})
+
+    @contextmanager
+    def context(self, context):
+        """Sets the context that Selenium commands are running in using
+        a `with` statement. The state of the context on the server is
+        saved before entering the block, and restored upon exiting it.
+
+        :param context: Context, may be one of the class properties
+            `CONTEXT_CHROME` or `CONTEXT_CONTENT`.
+
+        Usage example::
+
+            with selenium.context(selenium.CONTEXT_CHROME):
+                # chrome scope
+                ... do stuff ...
+        """
+        initial_context = self.execute('GET_CONTEXT').pop('value')
+        self.set_context(context)
+        try:
+            yield
+        finally:
+            self.set_context(initial_context)
